Add initial version of mapping function which allocates with the bump allocator.
