+++ crates/arch/x86/src/structures/paging/page_table.rs
@@ -101,44 +101,3 @@ impl PageTable {
         EntryIndex::OutOfEntries
     }
 }
-
-#[extend::ext]
-pub impl VirtualAddress {
-    #[cfg(target_arch = "x86_64")]
-    fn walk(&self) -> impl Iterator<Item = NonNull<PageTableEntry>> {
-        let mut table = Some(PageTable::current_table());
-        let mut level = Some(PageTableLevel::PML4);
-        ::core::iter::from_fn(move || {
-            let entry =
-                unsafe { &table?.as_ref().entries[self.index_of(level?)] };
-
-            if entry.get_flags().is_present() {
-                table = entry.mapped_table().ok();
-                level = level?.next();
-                Some(NonNull::from_ref(entry))
-            } else {
-                None
-            }
-        })
-    }
-
-    #[cfg(target_arch = "x86_64")]
-    fn map(
-        &self,
-        address: PhysicalAddress,
-        flags: Option<PageEntryFlags>,
-        page_size: PageSize,
-    ) -> Result<(), MappingError> {
-        let mut entry = self
-            .walk()
-            .nth(page_size.mapping_table() as usize)
-            .ok_or(MappingError::TableDoesNotExist)?;
-        unsafe {
-            entry.as_mut().map(
-                address,
-                flags.unwrap_or(PageEntryFlags::regular_page_flags()),
-            )
-        }
-        Ok(())
-    }
-}
