+++ crates/libk/Cargo.toml
@@ -5,3 +5,6 @@ edition = "2024"
 
 [dependencies]
 common = { path = "../common" }
+bump = { path = "../memory/bump" }
+x86 = { path = "../arch/x86" }
+extend = "1.2.0"
