+++ crates/libk/src/lib.rs
@@ -1,33 +1,4 @@
 #![no_std]
-
-unsafe extern "Rust" {
-    #[allow(unused)]
-    pub unsafe fn kprint(args: ::core::fmt::Arguments<'_>);
-}
-
-/// Prints formatted text to the VGA display without a
-/// newline.
-///
-/// # Parameters
-/// - `$fmt`: The format string.
-/// - `$arg`: Optional arguments to interpolate into the format string.
-/// - `color = $color`: Optional named parameter to change the VGA text
-///   color for this print.
-#[macro_export]
-macro_rules! print {
-    // Case 1: Standard print with optional arguments.
-    ($fmt:expr $(, $arg:expr)* $(;)?) => {{
-        unsafe { $crate::kprint(format_args!($fmt, $($arg,)*)) }
-    }};
-}
-
-/// Prints formatted text followed by a newline to the VGA
-/// display. Same as the [`print!`] macro just with a `\n`
-/// attached to the end
-#[macro_export]
-macro_rules! println {
-    // Case 1: Standard println with optional arguments.
-    ($fmt:expr $(, $arg:expr)* $(;)?) => {
-        $crate::print!(concat!($fmt, "\n") $(, $arg)*)
-    };
-}
+#![feature(allocator_api)]
+pub mod alloc;
+pub mod fmt;
