+++ crates/memory/page/Cargo.toml
@@ -6,10 +6,12 @@ edition = "2024"
 [dependencies]
 common = { path = "../../common" }
 x86 = { path = "../../arch/x86" }
-buddy = { path = "../buddy" }
+# buddy = { path = "../buddy" }
 num_enum = { git = "https://github.com/sagi21805/num_enum.git", rev = "2e53bd356ae93c537f3345f38679617bb977633c", default-features = false, features = [
     "complex-expressions",
 ] }
 macros = { path = "../../macros" }
+libk = { path = "../../libk" }
 strum_macros = { version = "0.27", default-features = false }
 strum = { version = "0.27", default-features = false }
+extend = "1.2.0"
