+++ crates/memory/page/src/lib.rs
@@ -1,39 +1,40 @@
 #![no_std]
 
+// pub mod arena;
 pub mod map;
-pub mod meta;
+// pub mod meta;
 
-use crate::meta::PageMeta;
-use buddy::meta::{BuddyBlock, BuddyMeta, Regular};
-use core::{cell::Ref, marker::PhantomData, ptr::NonNull};
+// use crate::meta::PageMeta;
+// use buddy::meta::{BuddyBlock, BuddyMeta, Regular};
+// use core::{cell::Ref, marker::PhantomData, ptr::NonNull};
 
-pub struct Page {
-    pub meta: PageMeta,
-}
+// pub struct Page {
+//     pub meta: PageMeta,
+// }
 
-impl BuddyBlock for Page {
-    #[inline]
-    fn from_meta(
-        meta: core::ptr::NonNull<BuddyMeta<Regular>>,
-    ) -> core::ptr::NonNull<Self> {
-        let offset = core::mem::offset_of!(Page, meta.buddy);
-        unsafe {
-            NonNull::new_unchecked(
-                meta.as_ptr().cast::<u8>().sub(offset).cast::<Self>(),
-            )
-        }
-    }
+// impl BuddyBlock for Page {
+//     #[inline]
+//     fn from_meta(
+//         meta: core::ptr::NonNull<BuddyMeta<Regular>>,
+//     ) -> core::ptr::NonNull<Self> {
+//         let offset = core::mem::offset_of!(Page, meta.buddy);
+//         unsafe {
+//             NonNull::new_unchecked(
+//                 meta.as_ptr().cast::<u8>().sub(offset).cast::<Self>(),
+//             )
+//         }
+//     }
 
-    #[inline]
-    fn meta(&self) -> &BuddyMeta<Regular> {
-        unsafe { &self.meta.buddy.regular }
-    }
+//     #[inline]
+//     fn meta(&self) -> &BuddyMeta<Regular> {
+//         unsafe { &self.meta.buddy.regular }
+//     }
 
-    #[inline]
-    fn meta_mut(&mut self) -> &mut BuddyMeta<Regular> {
-        unsafe { &mut self.meta.buddy.regular }
-    }
-}
+//     #[inline]
+//     fn meta_mut(&mut self) -> &mut BuddyMeta<Regular> {
+//         unsafe { &mut self.meta.buddy.regular }
+//     }
+// }
 
 // impl AssignSlab for NonNull<Page<()>> {
 //     type Target<Unassigned: Slab> = NonNull<Page<Unassigned>>;
