+++ kernel/Cargo.toml
@@ -22,3 +22,4 @@ nonmax = { version = "0.5.5", default-features = false }
 # pci = { path = "../crates/drivers/pci" }
 vga_display = { path = "../crates/drivers/vga-display" }
 bump = { path = "../crates/memory/bump"}
+libk = { path = "../crates/libk" }
