+++ kernel/src/main.rs
@@ -2,6 +2,7 @@
 #![no_main]
 #![feature(ptr_alignment_type)]
 #![feature(allocator_api)]
+#![allow(static_mut_refs)]
 
 use core::{
     alloc::{Allocator, Layout},
@@ -10,9 +11,14 @@ use core::{
 };
 
 use bump::BumpAllocator;
-use common::constants::{
-    MEMORY_MAP_LENGTH, MEMORY_MAP_OFFSET, PARSED_MEMORY_MAP,
-    REGULAR_PAGE_ALIGNMENT, REGULAR_PAGE_SIZE,
+use common::{
+    address_types::{Address, PhysicalAddress, VirtualAddress},
+    constants::{
+        MEMORY_MAP_LENGTH, MEMORY_MAP_OFFSET, MiB, PARSED_MEMORY_MAP,
+        PHYSICAL_MEMORY_OFFSET, REGULAR_PAGE_ALIGNMENT, REGULAR_PAGE_SIZE,
+    },
+    enums::PageSize,
+    late_init::LateInit,
 };
 // use keyboard::ps2_keyboard::Keyboard;
 use vga_display::{eprintln, okprintln, println};
@@ -21,6 +27,10 @@ use x86::{
     memory_map::{MemoryMap, MemoryRegion, MemoryRegionExtended},
     structures::paging::{PageTable, PageTableEntry},
 };
+
+use libk::alloc::{BUMP_ALLOCATOR, VirtualAddressExt};
+
+static mut MMAP: LateInit<MemoryMap> = LateInit::uninit();
 
 #[unsafe(no_mangle)]
 #[unsafe(link_section = ".start")]
@@ -44,11 +54,18 @@ pub unsafe extern "C" fn _start() -> ! {
         )
     };
 
-    let mmap = MemoryMap::parse_map(raw, buf).unwrap();
+    let mmap = MMAP.write(MemoryMap::parse_map(raw, buf).unwrap());
 
     println!("{}", mmap);
 
-    let bump = BumpAllocator::new(&mmap);
+    let bump =
+        BUMP_ALLOCATOR.write(BumpAllocator::new(MMAP.assume_init_ref()));
+
+    let v =
+        VirtualAddress::new_unchecked(PHYSICAL_MEMORY_OFFSET + 0x10000);
+    let p = PhysicalAddress::new_unchecked(6 * MiB);
+
+    v.map(p, None, PageSize::Big, &bump);
 
     // okprintln!("Obtained Memory Map");
     // println!("{}", MemoryMap(parsed_memory_map!()));
